Issue 7: Optionally insert Signed-off-by line in the commit message
The line is inserted if there are no previous commit message. A module-wide
option makes it possible to configure a default value while it is also possible
to configure the repository specific option via the properties UI. The option
is stored in .git/config using the name nbgit.signoff.

This commit polishes the layout of the options dialog to get better resize behavior
and spacing between form items.

Signed-off-by: Jonas Fonseca <fonseca@diku.dk>