Issue 16: Add UI for user defined actions
Called "Custom Actions", user are now able to create their own actions.
Each action definition consists of an action name, path to binary or script,
arguments (which can contain replacable tokens) and a handful of options
controlling whether to show output or not.

The general goal is to provide a temporary solution to some of the missing
pieces. For example a use can define a "Gitk All Branches" action to run
"gitk --all" in the repository root.

Signed-off-by: Jonas Fonseca <fonseca@diku.dk>