Rewrite the Excludes code
This moves pattern matching and reading of .gitignore files out of
Excludes into the new ExcludeCache class, which holds all exclude
patterns read from a repository. This makes it easier to test the
in-memory representation of .gitignore and .git/info/exclude files.
One test is introduced, which builds a "live" repository and checks
if the cache excludes and includes the correct files.

Most importantly it also changes the algorithm for matching a file to
start from the work directory root and iterate down into the
subdirectories ending with the file being queried.

Signed-off-by: Jonas Fonseca <fonseca@diku.dk>